requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
httpx>=0.27.0
youtube-transcript-api>=0.6.2
yt-dlp>=2024.1.1
chromadb>=0.5.0
python-dotenv>=1.0.0
pydantic>=2.0.0
pytest>=8.0.0
